backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.9
Pillow==10.4.0
brother_ql==0.9.4
zeroconf==0.131.0
gspread==6.1.2
google-auth==2.34.0
zxing-cpp
